// src/components/burger-ingredients/dto.ts
export interface IBurgerIngredient {
  _id: string;
  name: string;
  type: string;
  proteins: number;
  fat: number;
  carbohydrates: number;
  calories: number;
  price: number;
  image: string;
  image_mobile: string;
  image_large: string;
  __v?: number;
}

export interface IBurgerIngredientResponse {
  data: IBurgerIngredient[];
  status: string;
}